Remove ability to hide turret from armed variant
