+++ addons/arcadian/base/config_animationsources_armed.hpp
@@ -1,0 +1,20 @@
+class AnimationSources: AnimationSources {
+    class Revolving {
+        source = "revolving";
+        weapon = QCLASS(LMG_Minigun_SUV);
+    };
+    class CloseCover {
+        source = "user";
+        initPhase = 0;
+        animPeriod = 1.2;
+    };
+    class rearseats_source {
+        source = "user";
+        animPeriod = 1;
+        initPhase = 1;
+        displayName = "Fold Rear Seats";
+        lockCargo[] = {4,5};
+        lockCargoAnimationPhase = 1;
+        useSource = 1;
+    };
+};
