+++ addons/arcadian/model.cfg
@@ -531,6 +531,70 @@ class CfgModels {
                 maxValue = 0;
                 hideValue = "0.2";
             };
+            class reardoor_2_source: Rotation {
+                selection = "reardoor_2";
+                source = reardoor_2_source;
+                axis = "reardoor_axis";
+                memory = 1;
+                minValue = 0;
+                maxValue = 1;
+                angle0 = 0;
+                angle1 = rad 80;
+            };
+            class ram_hide_source {
+                type="hide";
+                source="ram_hide_source";
+                selection="ram_hide";
+                maxValue=1;
+                minValue=0;
+                hideValue=1;
+            };
+            class antenna_hide_source {
+                type="hide";
+                source="antenna_hide_source";
+                selection="antenna_hide";
+                maxValue=1;
+                minValue=0;
+                hideValue=1;
+            };
+            class roofbar_hide_source {
+                type="hide";
+                source="roofbar_hide_source";
+                selection="roofbar_hide";
+                maxValue=1;
+                minValue=0;
+                hideValue=1;
+            };
+            class roofbarexternal_hide_source: roofbar_hide_source {
+                type="hide";
+                source="roofbar_hide_source";
+                selection="roofbarexternal_hide";
+                maxValue=1;
+                minValue=0;
+                unhideValue=1;
+                hideValue=0;
+            };
+        };
+    };
+    class Arcadian_Armed: Arcadian {
+        skeletonName = "Arcadian_Armed";
+        class Animations: Animations {
+            class CloseCover1 {
+                selection = "poklop1";
+                type = "rotation";
+                source = "close_cover_source"; //CloseCover1
+                axis = "poklop1_axis";
+                memory = 1;
+                minValue = 0;
+                maxValue = 1;
+                angle1 = 0;
+                angle0 = rad 90;
+            };
+            class CloseCover2: CloseCover1 {
+                selection = "poklop2";
+                source = "close_cover_source"; //CloseCover2
+                axis = "poklop2_axis";
+            };
             class MainTurret {
                 type = "rotation";
                 source = "mainTurret";
@@ -587,16 +651,6 @@ class CfgModels {
                 maxValue = 1;
                 angle0 = rad -101.8;
             };
-            class reardoor_2_source: Rotation {
-                selection = "reardoor_2";
-                source = reardoor_2_source;
-                axis = "reardoor_axis";
-                memory = 1;
-                minValue = 0;
-                maxValue = 1;
-                angle0 = 0;
-                angle1 = rad 80;
-            };
             class rearseats_source {
                 selection = "rear_seats";
                 type = "rotation";
@@ -618,71 +672,6 @@ class CfgModels {
                 maxValue = 1;
                 angle0 = 0;
                 angle1 = rad -20;
-            };
-            class ram_hide_source {
-                type="hide";
-                source="ram_hide_source";
-                selection="ram_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
-            };
-            class antenna_hide_source {
-                type="hide";
-                source="antenna_hide_source";
-                selection="antenna_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
-            };
-            class roofbar_hide_source {
-                type="hide";
-                source="roofbar_hide_source";
-                selection="roofbar_hide";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
-            };
-            class roofbarexternal_hide_source: roofbar_hide_source {
-                type="hide";
-                source="roofbar_hide_source";
-                selection="roofbarexternal_hide";
-                maxValue=1;
-                minValue=0;
-                unhideValue=1;
-                hideValue=0;
-            };
-            class gun_hide_source {
-                type="hide";
-                source="gun_hide_source";
-                selection="OtocVez";
-                maxValue=1;
-                minValue=0;
-                hideValue=1;
-            };
-            class damageHidePolice: damageHide {
-                selection = "lightbar";
-            };
-        };
-    };
-    class Arcadian_Armed: Arcadian {
-        skeletonName = "Arcadian_Armed";
-        class Animations: Animations {
-            class CloseCover1 {
-                selection = "poklop1";
-                type = "rotation";
-                source = "close_cover_source"; //CloseCover1
-                axis = "poklop1_axis";
-                memory = 1;
-                minValue = 0;
-                maxValue = 1;
-                angle1 = 0;
-                angle0 = rad 90;
-            };
-            class CloseCover2: CloseCover1 {
-                selection = "poklop2";
-                source = "close_cover_source"; //CloseCover2
-                axis = "poklop2_axis";
             };
         };
     };
